+++ src/FIAP.TechChallenge.LambdaPagamento.Application/Models/Response/PagamentoResponse.cs
@@ -5,7 +5,7 @@ namespace FIAP.TechChallenge.LambdaPagamento.Application.Models.Response
     public class PagamentoResponse
     {
         public Guid Id { get; set; }
-        public decimal ValorTotal { get; set; }
+        public double ValorTotal { get; set; }
         public string StatusPagamento { get; set; }
 
         public string QrCode { get; set; }
